Addressing comments.
