+++ kythe/go/test/testutil/BUILD
@@ -5,5 +5,8 @@ package(default_visibility = ["//kythe:default_visibility"])
 go_library(
     name = "testutil",
     srcs = ["testutil.go"],
-    deps = ["@com_github_golang_protobuf//proto:go_default_library"],
+    deps = [
+        "@com_github_golang_protobuf//proto:go_default_library",
+        "@com_github_google_go_cmp//cmp:go_default_library",
+    ],
 )
